# eth/tracers/logger: remove unnecessary comparisons in accessList.equal