@karalabe 2022-04-14
Please remove the benchmark file. It's kind of to be expected that half the work results in 50% less execution time. Otherwise LGTM