@odeke-em 2022-04-21
Thanks for the reviews @karalabe @rjl493456442 @fjl! I've removed the benchmark file. However, personally I'd recommend keeping around benchmarks as there are tools such as our very own https://bencher.orijtech.com/ for continous profiling that can help pin point performance regressions, but that's something for the future.

Please take a look at the PR again, I've removed the benchmark file.