@fjl 2022-04-21
Thanks for the PR! We didn't want the benchmarks because we have A LOT of benchmarks already, and this function is not performance-critical.